+++ tests/testthat/test-expect-equality.R
@@ -5,6 +5,11 @@ test_that("basically principles of equality hold", {
   expect_failure(expect_equal(1, 2))
 })
 
+test_that("default labels use unquoting", {
+  x <- 2
+  expect_failure(expect_equal(1, !!x), "1 not equal to 2")
+})
+
 test_that("expect_equivalent ignores attributes for integers", {
   x1 <- 1:10
   x2 <- structure(x1, a = 1, b = 2)